requirements/test.txt
# Test dependencies go here.
-r base.txt
bok-choy==0.7.0
coverage==4.4.1
ddt==1.1.1
django-dynamic-fixture==1.9.5
django-nose==1.4.4
elasticsearch==2.4.1
unittest2==1.1.0
httpretty==0.8.14
mock==1.3.0
nose-exclude==0.5.0
nose-ignore-docstring==0.2
nose==1.3.7
pep257==0.7.0
pep8==1.7.0
pylint==1.7.1
logilab-common==1.4.0
selenium>=2.44.0,<3.0.0
sure==1.4.6
testfixtures==4.14.3
